backend_api_python/app/services/backtestsys_plugin/data_io/requirements.txt
ccxt>=4.0
pandas>=2.0
pyyaml>=6.0
pyarrow>=15.0
